refactor: fix lazy loading for limit orders and twap
