+++ apps/cowswap-frontend/src/common/utils/isUnsupportedTokenInQuote.ts
@@ -1,6 +1,5 @@
-import { TradeQuoteState } from 'modules/tradeQuote'
-
-import { QuoteApiErrorCodes } from 'api/cowProtocol/errors/QuoteError'
+import { QuoteApiErrorCodes } from '../../api/cowProtocol/errors/QuoteError'
+import { TradeQuoteState } from '../../modules/tradeQuote'
 
 export function isUnsupportedTokenInQuote(state: TradeQuoteState): boolean {
   return state.error?.type === QuoteApiErrorCodes.UnsupportedToken
