+++ apps/cowswap-frontend/src/common/utils/parsePrice.ts
