+++ apps/cowswap-frontend/src/modules/advancedOrders/containers/AdvancedOrdersWidget/index.tsx
@@ -5,9 +5,6 @@ import { isSellOrder } from '@cowprotocol/common-utils'
 
 import { Field } from 'legacy/state/types'
 
-import { useAdvancedOrdersActions } from 'modules/advancedOrders/hooks/useAdvancedOrdersActions'
-import { useAdvancedOrdersDerivedState } from 'modules/advancedOrders/hooks/useAdvancedOrdersDerivedState'
-import { advancedOrdersSettingsAtom } from 'modules/advancedOrders/state/advancedOrdersSettingsAtom'
 import { TradeWidget, TradeWidgetSlots, useReceiveAmountInfo, useTradePriceImpact } from 'modules/trade'
 import { BulletListItem, UnlockWidgetScreen } from 'modules/trade/pure/UnlockWidgetScreen'
 import { useTradeQuote } from 'modules/tradeQuote'
@@ -15,7 +12,10 @@ import { TWAP_LEARN_MORE_LINK } from 'modules/twap/const'
 
 import { CurrencyInfo } from 'common/pure/CurrencyInputPanel/types'
 
+import { useAdvancedOrdersActions } from '../../hooks/useAdvancedOrdersActions'
+import { useAdvancedOrdersDerivedState } from '../../hooks/useAdvancedOrdersDerivedState'
 import { useUpdateAdvancedOrdersRawState } from '../../hooks/useAdvancedOrdersRawState'
+import { advancedOrdersSettingsAtom } from '../../state/advancedOrdersSettingsAtom'
 import { AdvancedOrdersSettings } from '../AdvancedOrdersSettings'
 
 const TWAP_BULLET_LIST_CONTENT: BulletListItem[] = [
