+++ apps/cowswap-frontend/src/modules/advancedOrders/pure/Settings/index.tsx
@@ -1,5 +1,6 @@
-import { AdvancedOrdersSettingsState } from 'modules/advancedOrders/state/advancedOrdersSettingsAtom'
 import { SettingsBox, SettingsContainer, SettingsTitle } from 'modules/trade/pure/Settings'
+
+import { AdvancedOrdersSettingsState } from '../../state/advancedOrdersSettingsAtom'
 
 export interface SettingsProps {
   state: AdvancedOrdersSettingsState
