+++ apps/cowswap-frontend/src/modules/trade/state/derivedTradeStateAtom.ts
@@ -1,9 +1,9 @@
 import { atom } from 'jotai'
 
-import { advancedOrdersDerivedStateAtom } from 'modules/advancedOrders'
-import { limitOrdersDerivedStateAtom } from 'modules/limitOrders'
-import { swapDerivedStateAtom } from 'modules/swap2'
-import { yieldDerivedStateAtom } from 'modules/yield'
+import { advancedOrdersDerivedStateAtom } from 'modules/advancedOrders/state/advancedOrdersAtom'
+import { limitOrdersDerivedStateAtom } from 'modules/limitOrders/state/limitOrdersRawStateAtom'
+import { swapDerivedStateAtom } from 'modules/swap2/state/swapRawStateAtom'
+import { yieldDerivedStateAtom } from 'modules/yield/state/yieldRawStateAtom'
 
 import { tradeTypeAtom } from './tradeTypeAtom'
